shift from checker to fixed output
As we can't host VPS for free, our only option to use checker would be to store it in DB. If we store the source file, we will have to compile it everytime which will slow down the verdict process. Storing binary executable files in DB is a bad practice, also takes much more space. So we are switching from using checker to using fixed output for every input.
Also refactored a little bit.